requirements.txt
numpy==1.16.0
Pillow==5.4.1
pypng==0.0.19
PyQRCode==1.2.1
qrtools==0.0.2
scipy==1.2.0
